help messeges cleaned, refactored

+++ source/commands/help.js
@@ -1,11 +1,12 @@
+const { Util } = require("discord.js");
 const { prefix } = require("../../config.json");
+const CommandTypeEnum = require("../tools/commandTypeEnum.js");
 
 module.exports = {
     name: "help",
     description: "List all of my commands or info about a specific command.",
-    aliases: ["commands"],
-    usage: " [command name]",
-    cooldown: 5,
+    usage: "*optional:* [command name]",
+    type: CommandTypeEnum.UTILITY,
     execute(message, args, _bot) {
         const data = [];
         const { commands } = message.client;
@@ -13,10 +14,47 @@ module.exports = {
         console.log(message.author.username + " requested help");
 
         if (!args.length) {
-            data.push("Here's a list of all my commands:");
-            data.push(commands.map((command) => command.name).join(", "));
+            data.push("***Here's a list of all my commands:***");
+
+            let musicCommandsList = [];
+            let soundEffectCommandsList = [];
+            let utilityCommandsList = [];
+            let otherCommandsList = [];
+
+            commands
+                .map((cmd) => cmd.name)
+                .forEach((element) => {
+                    let cmd = commands.get(element);
+                    switch (cmd.type) {
+                        case CommandTypeEnum.MUSIC:
+                            musicCommandsList.push(cmd.name);
+                            break;
+                        case CommandTypeEnum.SOUNDEFFECT:
+                            soundEffectCommandsList.push(cmd.name);
+                            break;
+                        case CommandTypeEnum.UTILITY:
+                            utilityCommandsList.push(cmd.name);
+                            break;
+                        case CommandTypeEnum.OTHER:
+                            otherCommandsList.push(cmd.name);
+                            break;
+                    }
+                });
+
+            data.push("\n***Music Commands***");
+            data.push(musicCommandsList.join(", "));
+
+            data.push("\n***SoundEffect Commands***");
+            data.push(soundEffectCommandsList.join(", "));
+
+            data.push("\n***Utility Commands ***");
+            data.push(utilityCommandsList.join(", "));
+
+            data.push("\n***Other Commands***");
+            data.push(otherCommandsList.join(", "));
+
             data.push(
-                `\nYou can send \`${prefix}help [command name]\` to get info on a specific command!`
+                `\n***You can send*** \`${prefix}help [command name]\` ***to get info on a specific command!***`
             );
 
             return message.author
@@ -38,9 +76,7 @@ module.exports = {
 
         const name = args[0].toLowerCase();
 
-        const command =
-            commands.get(name) ||
-            commands.find((c) => c.aliases && c.aliases.includes(name));
+        const command = commands.get(name);
 
         if (!command) {
             return message.reply("that's not a valid command!");
@@ -55,10 +91,9 @@ module.exports = {
         if (command.usage)
             data.push(`**Usage:** ${prefix}${command.name} ${command.usage}`);
 
-        data.push(`**Cooldown:** ${command.cooldown || 3} second(s)`);
-
         message.channel.send(data, { split: true });
 
+        if (message.channel.type === "dm") return;
         message.delete();
     },
 };
